KFSPTS-10642: Updates from PR feedback
