+++ src/main/java/edu/cornell/kfs/pdp/batch/service/CuProcessPdpCancelPaidService.java
@@ -2,15 +2,13 @@ package edu.cornell.kfs.pdp.batch.service;
 
 import java.sql.Date;
 
+import org.kuali.kfs.pdp.batch.service.ProcessPdpCancelPaidService;
 import org.kuali.kfs.pdp.businessobject.PaymentDetail;
 
-/**
- * Helper interface for processing canceled/paid PDP payment details individually,
- * allowing each one to be processed in a separate transaction.
- */
-public interface ProcessPdpCancelPaidHelperService {
+public interface CuProcessPdpCancelPaidService extends ProcessPdpCancelPaidService {
 
     void processPdpCancel(PaymentDetail paymentDetail, Date processDate);
 
     void processPdpPaid(PaymentDetail paymentDetail, Date processDate);
+
 }
